lab1/backend/requirements.txt
djangorestframework
markdown
django-filter
psycopg[binary,pool]
django-extensions
python-dotenv